migrate from INTEGER to INT64

+++ src/core/ids/system.ts
@@ -17,7 +17,7 @@ export const BLOCKS = Id('beaba5cba67741a8b35377030613fc70');
 
 /** Data types */
 export const BOOLEAN = Id('7aa4792eeacd41868272fa7fc18298ac');
-export const INTEGER = Id('149fd752d9d04f80820d1d942eea7841');
+export const INT64 = Id('149fd752d9d04f80820d1d942eea7841');
 export const FLOAT64 = Id('9b597aaec31c46c88565a370da0c2a65');
 export const DECIMAL = Id('a3288c22a0564f6fb409fbcccb2c118c');
 export const TEXT = Id('9edb6fcce4544aa5861139d7f024c010');
